refactor(frontdesk): keep the rearm-watch join in its own file so autosync.go stays under its size ceiling
The rearm-watch join added on this branch pushed autosync.go and its test past
the counts the size gate has recorded for them, and an allowlisted file may
shrink but never grow. The watcher pair and its test move to files of their own:
startRearmWatch and watchRearm to autosync_rearm.go, TestRearmWatchStopJoinsThe
Watcher to autosync_rearm_test.go. Nothing about the code changes, so the ratchet
tightens on autosync.go rather than loosening.
