+++ scripts/ci/size-allowlist.txt
@@ -20,7 +20,7 @@ internal/api/logs.go	809
 internal/api/stats.go	843
 internal/failover/circuitbreaker_test.go	2056
 internal/failover/failover.go	1091
-internal/frontdesk/autosync.go	1196
+internal/frontdesk/autosync.go	1171
 internal/frontdesk/autosync_test.go	4254
 internal/frontdesk/poller.go	930
 internal/model/model.go	983
